refactor DocxFile
